rename multiroot managers to infomanagers
