+++ src/infoManagers/endpointManager.ts
